Refactor getPageSize utility from NetworkEnvironmentConfiguration to ConfigurationParserUtils
